@@ -1,18 +1,17 @@
 name: grock
 description: Snackbars and tools.
-version: 0.0.4+2
+version: 0.0.5
 # author: Ahmet Özberk <ahmet.ozberk.mrs@gmail.com>
 homepage: https://github.com/ahmet-ozberk/grock
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
   page_transition: ^2.0.5
-  provider: ^6.0.2
 
 dev_dependencies:
   flutter_test:
